Rename richtextblock to slate.
