+++ packages/editor/src/client/blocks/linkPageBreakBlock.tsx
@@ -3,7 +3,7 @@ import React, {useRef, useEffect, useState, useCallback} from 'react'
 import {Drawer, IconButton, Icon, Input} from 'rsuite'
 import {BlockProps} from '../atoms/blockList'
 import {LinkPageBreakBlockValue, RichTextBlockValue} from './types'
-import {createDefaultValue, RichTextBlock} from './richTextBlock/richTextBlock'
+import {createDefaultValue, RichTextBlock} from './richTextBlock/slate'
 import {ImageSelectPanel} from '../panel/imageSelectPanel'
 import {ImagedEditPanel} from '../panel/imageEditPanel'
 import {isFunctionalUpdate} from '@karma.run/react'
