+++ packages/editor/src/client/blocks/richTextBlock/slate.tsx
